requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
gspread_asyncio==1.5.0
oauth2client==4.1.3
python-dotenv